ocr/requirements.txt
google-cloud-vision
python-docx
docx2pdf